pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claimverify"
version = "0.1.0"
description = "Scientific Claim Verification with RAG - UZH FS 2026"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "torch>=2.1",
    "transformers>=4.40",
    "datasets>=2.18",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "accelerate>=0.28",

    # Retrieval
    "ir-datasets>=0.5",
    "nltk>=3.8",

    # Evaluation
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.2",

    # Config & CLI
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "rich>=13.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "ipykernel>=6.29",
]
train = [
    "peft>=0.10",
    "trl>=0.8",
    "bitsandbytes>=0.43",
    "wandb>=0.16",
]

[project.scripts]
claimverify = "claimverify.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
